fix(types): resolve mypy errors surfaced by the diagnostics->identify import
The new diagnostics.face_status -> identify.insightface_root import pulled
identify.py and service.py into CI's mypy scope (autoptz/engine/runtime/), which
surfaced 3 pre-existing latent type errors. Both are behavior-neutral:
- identify._yaw_ratio: wrap the Any-typed division in float() to match the
  declared float|None return.
- service.enroll: type the conditional id kwargs as dict[str, Any] so the **
  unpack doesn't make mypy infer every IdentityRecord field as str.

+++ autoptz/engine/pipeline/identify.py
@@ -92,7 +92,7 @@ class FaceObservation:
         inter_ocular = ((rx - lx) ** 2 + (ry - ly) ** 2) ** 0.5
         if inter_ocular <= 1e-3:
             return None
-        return abs(nx - eye_mid_x) / inter_ocular
+        return float(abs(nx - eye_mid_x) / inter_ocular)
 
 
 @dataclass(frozen=True)
